fix: temporal dataloader, with added path sanitize
