+++ data/unaligned_labeled_mask_online_dataset.py
@@ -10,7 +10,7 @@ from PIL import Image
 
 from data.base_dataset import BaseDataset, get_transform, get_transform_seg
 from data.image_folder import make_dataset, make_dataset_path, make_labeled_path_dataset
-from data.online_creation import crop_image, sanitize_paths, write_paths_file
+from data.online_creation import crop_image
 
 
 class UnalignedLabeledMaskOnlineDataset(BaseDataset):
@@ -76,81 +76,6 @@ class UnalignedLabeledMaskOnlineDataset(BaseDataset):
         self.semantic_nclasses = self.opt.f_s_semantic_nclasses
 
         self.header = ["img", "mask"]
-
-    def sanitize(self):
-        paths_sanitized_train_A = os.path.join(
-            self.sv_dir, "paths_sanitized_train_A.txt"
-        )
-        if hasattr(self, "B_img_paths"):
-            paths_sanitized_train_B = os.path.join(
-                self.sv_dir, "paths_sanitized_train_B.txt"
-            )
-        if hasattr(self, "B_img_paths"):
-            train_sanitized_exist = os.path.exists(
-                paths_sanitized_train_A
-            ) and os.path.exists(paths_sanitized_train_B)
-        else:
-            train_sanitized_exist = os.path.exists(paths_sanitized_train_A)
-
-        if train_sanitized_exist:
-            self.A_img_paths, self.A_label_mask_paths = make_labeled_path_dataset(
-                self.sv_dir, "/paths_sanitized_train_A.txt"
-            )
-            if hasattr(self, "B_img_paths"):
-                self.B_img_paths, self.B_label_mask_paths = make_labeled_path_dataset(
-                    self.sv_dir, "/paths_sanitized_train_B.txt"
-                )
-        else:
-            print("--------------")
-            print("Sanitizing images and labels paths")
-            print("--- DOMAIN A ---")
-
-            self.A_img_paths, self.A_label_mask_paths = sanitize_paths(
-                self.A_img_paths,
-                self.A_label_mask_paths,
-                mask_delta=self.opt.data_online_creation_mask_delta_A,
-                mask_random_offset=self.opt.data_online_creation_mask_random_offset_A,
-                crop_delta=self.opt.data_online_creation_crop_delta_A,
-                mask_square=self.opt.data_online_creation_mask_square_A,
-                crop_dim=self.opt.data_online_creation_crop_size_A,
-                output_dim=self.opt.data_load_size,
-                max_dataset_size=self.opt.data_max_dataset_size,
-                context_pixels=self.opt.data_online_context_pixels,
-                load_size=self.opt.data_online_creation_load_size_A,
-                select_cat=self.opt.data_online_select_category,
-                data_relative_paths=self.opt.data_relative_paths,
-                data_root_path=self.opt.dataroot,
-            )
-            write_paths_file(
-                self.A_img_paths,
-                self.A_label_mask_paths,
-                paths_sanitized_train_A,
-            )
-
-            print("--- DOMAIN B ---")
-            if hasattr(self, "B_img_paths"):
-                self.B_img_paths, self.B_label_mask_paths = sanitize_paths(
-                    self.B_img_paths,
-                    self.B_label_mask_paths,
-                    mask_delta=self.opt.data_online_creation_mask_delta_B,
-                    mask_random_offset=self.opt.data_online_creation_mask_random_offset_B,
-                    crop_delta=self.opt.data_online_creation_crop_delta_B,
-                    mask_square=self.opt.data_online_creation_mask_square_B,
-                    crop_dim=self.opt.data_online_creation_crop_size_B,
-                    output_dim=self.opt.data_load_size,
-                    max_dataset_size=self.opt.data_max_dataset_size,
-                    context_pixels=self.opt.data_online_context_pixels,
-                    load_size=self.opt.data_online_creation_load_size_B,
-                    data_relative_paths=self.opt.data_relative_paths,
-                    data_root_path=self.opt.dataroot,
-                )
-                write_paths_file(
-                    self.B_img_paths,
-                    self.B_label_mask_paths,
-                    paths_sanitized_train_B,
-                )
-
-            print("--------------")
 
     def get_img(
         self,
